requirements.txt
tornado==5.1.1
tormysql==0.4.0
envparse==0.2.0
wtforms-tornado==0.0.2
tornado-jinja2==0.2.4
beautifulsoup4==4.6.3
nltk==3.4
pycrypto==2.6.1
coverage==4.5.2
tornado-httpclient-mock==0.2.3
